Changes to tool parenting.
Made InvalidDrop and ToolDeletion parent the tool to the backpack instead of character. If you parent the tool to a character it might not be equipped and some hacky behavior might occur. Also because the behavior for tools when .CanBeDropped is false and you press backspace is that the tool gets deselected but if something on the client turns on .CanBeDropped then the deselection behavior of the tool via Backspace is broken.