Update init.lua
On #6 I added a feature where I fixed a bug with network ownership. However it introduced a new bug which makes it so that if the network ownership cannot be set then the anti cheat will stay enabled. So this makes is to that if for example the rootpart is anchored or for some other reason doesn't have network ownership the anticheat will also be disabled.